# waiting-room/waiting_room_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Waiting room mock API",
    "version": "1.0.0",
    "description": ""
  },
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/simulateUsers": {
      "post": {
        "description": "",
        "operationId": "simulateUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "numberOfUsers": {
                    "type": "integer"
                  }
                }
              }
            },
            "application/xml": {
              "schema": {
                "type": "object",
                "properties": {
                  "numberOfUsers": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dequeueUsers": {
      "post": {
        "description": "",
        "operationId": "dequeueUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "numberOfUsers": {
                    "type": "integer"
                  }
                }
              }
            },
            "application/xml": {
              "schema": {
                "type": "object",
                "properties": {
                  "numberOfUsers": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/log/{lineNumber}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "lineNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}